pyproject.toml
[project]
name = "mbcp"
version = "0.1.0"
description = "A tool for Minecraft particle production"
authors = [
    {name = "snowykami", email = "snowykami@outlook.com"},
]
dependencies = [
    "pytest~=8.3.2",
    "numpy~=2.0.1",
    "liteyukibot>=6.3.9",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
